vd_lavc: expose mastering display side data reference peak
This greatly improves the result when decoding typical (ST.2084) HDR
content, since the job of tone mapping gets significantly easier when
you're only mapping from 1000 to 250, rather than 10000 to 250.

The difference is so drastic that we can now even reasonably use
`hdr-tone-mapping=linear` and get a very perceptually uniform result
that is only slightly darker than normal. (To compensate for the extra
dynamic range)

Due to weird implementation details, this only seems to be present on
keyframes (or something like that), so we have to cache the last seen
value for the frames in between.

Also, in some files the metadata is just completely broken /
nonsensical, so I decided to apply a simple heuristic to detect
completely broken metadata.

+++ video/decode/lavc.h
@@ -26,7 +26,7 @@ typedef struct lavc_ctx {
     bool hwdec_notified;
 
     // For HDR side-data caching
-    int cached_hdr_peak;
+    double cached_hdr_peak;
 
     struct mp_image **delay_queue;
     int num_delay_queue;
